Exclude unneeded filed from distribution
The `test` directory takes up a good chunk of disk space. By removing it, the distribution becomes downright tiny. This might seem excessive, but I'm thinking of including `marked-man` as a standard feature of something I work on, so having it be this much smaller would be lovely and much easier to justify.

```
4.0K    node_modules/marked-man/bin
12K     node_modules/marked-man/lib
4.0K    node_modules/marked-man/man
80K     node_modules/marked-man/test/man
76K     node_modules/marked-man/test/md
96K     node_modules/marked-man/test/out
256K    node_modules/marked-man/test
304K    node_modules/marked-man
```