Make the output reproducible
Whilst working on the Reproducible Builds effort [0], I noticed that marked-man
generates its output nondeterminstically if the value of SOURCE_DATE_EPOCH is
within X hours of the end of the UTC month and the build system's timezone
offset is greater than or equal to X hours.

You can see an example here:

  https://bugs.debian.org/1030724#5

The solution is to emit the output as UTC and not in the build system's
timezone.

I originally filed this in Debian as bug #1030724 [1].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/1030724